Polyline KiCAD export not divide into single lines
The polyline KiCAD export algorithm was dividing polylines into separate
single lines. This may be useful, but is not always what user wants.
The patch introduces a class property to choose whether polyline should
be divided on export or not. It also adds proper inclusion of
slot/unit index into the resulting lines.